+++ node.cpp
@@ -1,6 +1,6 @@
 #include "node.h"
 
-Node::Node(int nID, int idxID, float x, float y, float z, Node_t node_type)
+Node::Node(int nID, long idxID, float x, float y, float z, Node_t node_type)
 {
     m_nID = nID;        // hvgx ID
     m_idxID = idxID;    // m_bufferNode id
